+++ yash-syntax/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `SyntaxError::RedundantToken` variant
+- `parser::SyntaxError::RedundantToken` variant
+- `parser::lex::ParseOperatorError` struct
 
 ### Changed
 
@@ -55,7 +56,6 @@ platforms, where `RawFd` falls back to `i32`.
 - `parser::lex::ParseKeywordError`
 - `impl FromStr for parser::lex::Keyword`
 - `parser::lex::Operator::as_str`
-- `parser::lex::ParseOperatorError`
 - `parser::lex::TryFromOperatorError`
 
 ### Changed